requirements.txt
streamlit
openai
google-genai
python-dotenv